// frontend/src/components/types.ts
export interface DebugServerInfo {
  svid_chain: [
        {
            id: {
                trust_domain: string,
                path: string
            },
            expires_at: number,
            subject: string
        },
        {
            id: {
                trust_domain: string
            },
            expires_at: number,
            subject: string
        }
    ],
    uptime: number,
    federated_bundles_count: number
}

export interface TornjakServerInfo {
  // Plugins for tornjak server info
  plugins: {
    DataStore: string[];
    KeyManager: string[];
    NodeAttestor: string[];
    NodeResolver: string[];
    Notifier: string[];
  };
  trustDomain: string; // Trust domain of server
  verboseConfig: string; // More detailed metadata on server
}

export interface ServerInfo {
  trustDomain: string; // Trust domain of server
  nodeAttestorPlugin: string; // Node Attestor Plugin of server
}
// ...
export interface AccessToken {
  // From https://www.keycloak.org/docs-api/11.0/javadocs/org/keycloak/representations/AccessToken.html
  acr: string; // Authentication Context Class Reference
  "allowed-origins": string[]; // allowed origins
  auth_time: bigint; // authentication time
  azp: string; // authorized party - to whom this token is issued
  email: string; // email of user
  email_verified: boolean; // wheather email is verified or not
  exp: bigint; // expiration time
  family_name: string; // family name of user
  given_name: string; // given name of user
  iat: bigint; // time issued
  iss: string; // entity that created and signed the token
  jti: string; // 
  name: string; // name of user
  nonce: string; // string value used to associate a Client session with an ID Token
  preferred_username: string; // 
  realm_access: {"roles": string[]}; // realm attributes
  scope: string; // scope
  session_state: string; // session state
  sid: string; // unique session identifier of a user on a browser or a device
  sub: string; // to whom this entity refers to
  typ: string; // type of token
}